Moved files to results folder

+++ main.py
@@ -13,16 +13,14 @@ FFMPEG_BITRATE = "32k"
 
 client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
 
-
 USER_PROMPT = """Transcript: {}"""
 
 def load_prompt(filename: str) -> str:
     with open(filename, "r") as f:
         return f.read()
 
-def download_audio_from_youtube(url):
+def download_audio_from_youtube(url, title):
     """Downloads audio from the given YouTube URL and returns the filename."""
-
     filename = None
 
     def my_hook(d):
@@ -30,8 +28,9 @@ def download_audio_from_youtube(url):
         if d["status"] == "finished":
             filename = d["filename"]
 
+
     ydl_opts = {
-        "outtmpl": "%(title)s.%(ext)s",
+        "outtmpl": f"results/{title}/{title}.%(ext)s",
         "format": "worstaudio",
         "postprocessors": [
             {
@@ -73,14 +72,15 @@ def transcribe_audio(audio_filename):
         )
     return transcription
 
-def summarize_transcript(transcript, system_prompt, user_prompt_template):
-    summarize_prompt = user_prompt_template.format(transcript)
+def summarize_transcript(transcript):
+    system_prompt = load_prompt("system_prompt.txt")
+    user_prompt_template = USER_PROMPT.format(transcript)
 
     response = client.chat.completions.create(
-        model="gpt-4-0125-preview",
+        model="gpt-4o",
         messages=[
             {"role": "system", "content": system_prompt},
-            {"role": "user", "content": summarize_prompt},
+            {"role": "user", "content": user_prompt_template},
         ],
         temperature=0,
         top_p=1,
@@ -89,28 +89,40 @@ def summarize_transcript(transcript, system_prompt, user_prompt_template):
     )
     return response.choices[0].message.content
 
-def main(url: Union[str, None]):
+def main(url: Union[str, None], title: Union[str, None]):
     """Main function to parse arguments and orchestrate the summarization of a YouTube video."""
-    if not url:
+    if not url or not title:
         parser = argparse.ArgumentParser(description="Summarize YouTube videos.")
         parser.add_argument(
             "url", type=str, help="The URL of the YouTube video to summarize."
         )
+        parser.add_argument(
+            "title", type=str, help="The title to use for naming files."
+        )
         args = parser.parse_args()
         url = args.url.replace("\\", "")
+        title = args.title
 
     try:
-        audio_filename = download_audio_from_youtube(url)
-        convert_audio_to_mono(audio_filename)
-        transcript = transcribe_audio(audio_filename)
-        with open(f"{audio_filename}.txt", "w") as f:
-            f.write(transcript)
-        system_prompt = load_prompt("system_prompt.txt")
-        user_prompt = USER_PROMPT.format(transcript)
-        summary = summarize_transcript(transcript, system_prompt, user_prompt)
-        # save the summary to a file
-        with open(f"{audio_filename}_summary.md", "w") as f:
-            f.write(summary)
+        audio_filename = f"results/{title}/{title}"
+        if not os.path.exists(f"{audio_filename}.{AUDIO_FORMAT}"):
+            audio_filename = download_audio_from_youtube(url, title)
+            
+        if not os.path.exists(f"{audio_filename}_mono.{AUDIO_FORMAT}"):
+            convert_audio_to_mono(audio_filename)
+
+        if not os.path.exists(f"{audio_filename}.txt"):
+            transcript = transcribe_audio(audio_filename)
+            with open(f"{audio_filename}.txt", "w") as f:
+                f.write(transcript)
+        else:
+            with open(f"{audio_filename}.txt", "r") as f:
+                transcript = f.read()
+
+        if not os.path.exists(f"{audio_filename}_summary.md"):
+            summary = summarize_transcript(transcript)
+            with open(f"{audio_filename}_summary.md", "w") as f:
+                f.write(summary)
     finally:
         pass
         # Cleanup downloaded and processed files only if they exist
@@ -120,6 +132,7 @@ def main(url: Union[str, None]):
         #     os.remove(f"{audio_filename}_mono.{AUDIO_FORMAT}")
 
 url = "https://www.youtube.com/watch?v=BT6Aw6Q75Yg"
+title = "All Learning Algorithms Explained in 14 Minutes"
 
 if __name__ == "__main__":
-    main(url)
+    main(url, title)
